+++ src/index.d.ts
@@ -1,4 +1,5 @@
 
+export type JSONPointer = string // a string containing a JSONPointer like '/array/3/name'
 export type JSONPath = (string | number)[]
 
 export type JSONValue = string | number | boolean | null
@@ -8,36 +9,36 @@ export type JSONData = JSONObject | JSONArray | JSONValue
 
 export interface JSONPatchAdd {
   op: 'add'
-  path: string
+  path: JSONPointer
   value: JSONData
 }
 
 export interface JSONPatchRemove {
   op: 'remove'
-  path: string
+  path: JSONPointer
 }
 
 export interface JSONPatchReplace {
   op: 'replace'
-  path: string
+  path: JSONPointer
   value: JSONData
 }
 
 export interface JSONPatchCopy {
   op: 'copy'
-  path: string
-  from: string
+  path: JSONPointer
+  from: JSONPointer
 }
 
 export interface JSONPatchMove {
   op: 'move'
-  path: string
-  from: string
+  path: JSONPointer
+  from: JSONPointer
 }
 
 export interface JSONPatchTest {
   op: 'test'
-  path: string
+  path: JSONPointer
   value: JSONData
 }
 
@@ -106,8 +107,11 @@ export declare function immutableJSONPatch (json: JSONData, operations: JSONPatc
 export declare function revertJSONPatch (json: JSONData, operations: JSONPatchDocument) : JSONPatchDocument
 
 // utils
-export declare function parseJSONPointer (pointer: string) : JSONPath
-export declare function compileJSONPointer (path: JSONPath) : string
+export declare function parseJSONPointer (pointer: JSONPointer) : JSONPath
+export declare function compileJSONPointer (path: JSONPath) : JSONPointer
+export declare function compileJSONPointerProp (pathProp: string | number) : JSONPointer
+export declare function appendToJSONPointer (pointer: JSONPointer, pathProp: string | number) : JSONPointer
+export declare function startsWithJSONPointer (pointer: JSONPointer, searchPointer: JSONPointer) : boolean
 export declare function getIn(json: JSONData, path: JSONPath) : JSONData
 export declare function setIn(json: JSONData, path: JSONPath, value: JSONData, createPath?: boolean) : JSONData
 export declare function updateIn(json: JSONData, path: JSONPath, callback: (json: JSONData) => JSONData) : JSONData