+++ src/index.js
@@ -2,7 +2,13 @@ export { immutableJSONPatch } from './immutableJSONPatch.js'
 export { revertJSONPatch } from './revertJSONPatch.js'
 
 // utils
-export { compileJSONPointer, parseJSONPointer } from './jsonPointer.js'
+export {
+  compileJSONPointer,
+  compileJSONPointerProp,
+  parseJSONPointer,
+  startsWithJSONPointer,
+  appendToJSONPointer
+} from './jsonPointer.js'
 export {
   getIn,
   setIn,